Rename config edit operation for null to avoid YAML mangling
The YAML parse interpreted the enum value of Null as a keyword. That
means when it reads in config edits for a mod it will change the
operation to the default enum value of Override! This is not good.

Also changed the enum name for Default to avoid similar possible issues.