Add a component for library DLL files
This follows the same pattern as config edits and asset bundles. The
component shows up in the dropdown in the bottom right and the metadata
includeLibraries field is toggled depending on if the component is not
null.

Multiple DLL files can be included and they can be sourced from paths
outside the mod project folder.