Add namespaces to function selector
The add button on function lists shows a flyout to select the function to
add. This is the standard Unity object picker. If you have
implementations in different namespaces, sometimes Unity will show a
list with foldouts for the namespaces and other times it will show just
a flat list. It's inconsistent.

Where this becomes a problem is when you drop implementations from
several different mods into the User folder. When you're working on a
mod, you don't want to pick up the implementations from other mods
because they won't necessarily be available in the game. This means you
should see the different namespaces for the implementations so you can
pick functions from the correct namespace.

This commit replaces the Unity object picker on lists for interfaces
that have been marked with the TypeHinted attribute. These are the most
likely interfaces that modders will write implementations for. If it
turns out they also implement the unmarked interfaces then it's easy to
add the custom selector for lists of those interfaces.