Add capability to include extra files in exported mod
Some modders may want to include READMEs, LICENSEs, CHANGELOGs and other
pertinent files with their mods. This is similar to including library
DLLs except that the files are copied to the top-level of the mod
folder instead of a subfolder like Libraries. These top-level files are
ignored by the game but may include things like settings which the mod
itself may use.