Specify the exact version of the Editor to install in the readme
There have been a few people who've installed a different version of the Editor
and got a number of confusing errors when loading the SDK. They've had to pop
into Discord to get help. It was suggested that specifying the full exact
version of the Editor would reduce the chance of this happening.

Signed-off-by: EchKode <48565771+echkode@users.noreply.github.com>