Load hinted types from user DLLs
When a user-supplied DLL is added to Assets/Users, Unity Editor will pick it up
and populate various dropdowns with any function interface implementations it
finds. However, the types are not added to the YAML serializer so if one of
these user-supplied functions is selected in a dropdown for a database entry, it
won't be properly serialized to the backing YAML file.

Signed-off-by: EchKode <48565771+echkode@users.noreply.github.com>