Remove incorrect namespace on entity classes
The entity classes were mistakenly put in the PhantomBrigade namespace but in
the game assemblies they are not in that namespace. In fact, they are "bare"
names without any namespacing.

The namespace breaks user implementations of function interfaces classes because
Unity Editor won't be able to correctly resolve the references to entity classes
in user-supplied DLLs.

Signed-off-by: EchKode <48565771+echkode@users.noreply.github.com>