Fix failed type lookup (#30)
The API to get a type from a path changed to allow partial lookups by default.
In the original update to 2.0, I had corrected one place where I was using it
but missed a second.

Some file APIs in Mono also choke on paths with mixed separators. All path
arguments to file operations in the checksum logic are canonicalized with the
clean path API.

Signed-off-by: EchKode <48565771+echkode@users.noreply.github.com>